requirements/dev_windows.txt
amqp==5.0.6
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
base58==2.0.1
billiard==3.6.4.0
bitarray==1.2.2
black==21.5b1
bump2version==1.0.1
bumpversion==0.6.0
celery==5.1.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coloredlogs==15.0.1
ConfigArgParse==1.4
cors==0.9.7
cryptography==3.3.2
cytoolz==0.11.0
Deprecated==1.2.13
enforce-typing==1.0.0.post1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.0
future==0.18.2
gevent==20.12.1
geventhttpclient==1.4.4
greenlet==0.4.17
hexbytes==0.2.1
humanfriendly==10.0
idna==2.10
ImageHash==4.2.0
install==1.3.4
ipfshttpclient==0.7.0a1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
json-sempai==0.4.0
jsonschema==3.2.0
kombu==5.1.0
lazy-object-proxy==1.4.3
locust==1.4.4
lru-dict==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
multiaddr==0.0.9
mypy-extensions==0.4.3
netaddr==0.8.0
nltk==3.5
numpy==1.19.3
ocean-contracts==0.6.9
ocean-lib==0.8.4
pandas==1.2.0
parsimonious==0.8.1
pathspec==0.8.1
Pillow==8.1.2
prompt-toolkit==3.0.20
protobuf==3.14.0
psutil==5.8.0
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.10.4
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==21.0.0
pypiwin32==223
pyreadline3==3.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
qrcode==6.1
regex==2020.11.13
requests==2.25.0
requests-file==1.5.1
rlp==2.0.1
scipy==1.5.4
six==1.15.0
tldextract==3.1.0
toml==0.10.2
toolz==0.11.1
tqdm==4.56.0
urllib3==1.26.6
varint==1.0.2
videohash==1.0.6
vine==5.0.0
wcwidth==0.2.5
web3==5.19.0
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2021.6.6
zope.event==4.5.0
zope.interface==5.2.0
